+++ src/main/java/nextstep/member/ui/LoginController.java
@@ -1,0 +1,24 @@
+package nextstep.member.ui;
+
+import nextstep.member.application.AuthService;
+import nextstep.member.application.dto.TokenRequest;
+import nextstep.member.application.dto.TokenResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class LoginController {
+
+    private AuthService authService;
+
+    public LoginController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login/token")
+    public ResponseEntity<TokenResponse> createToken(@RequestBody TokenRequest request) {
+        return ResponseEntity.ok().body(authService.createToken(request));
+    }
+}
